fix(desktop): renderer 不再经 context-usage 引用 host-internal
Phase 2 在 settings-view 导入 context-usage 时连带 host-internal 与 storage（keyring），触发 Vite 预构建失败。将 parseModelContextLength 拆至 renderer 安全的 model-context-length 模块。
