+++ apps/desktop/src/lib/model-context-length.ts
@@ -1,0 +1,9 @@
+/** Renderer-safe: no host-internal / host/storage imports. */
+
+export function parseModelContextLength(value: unknown): number | undefined {
+  if (typeof value !== 'number' || !Number.isFinite(value) || !Number.isInteger(value) || value <= 0) {
+    return undefined;
+  }
+
+  return value;
+}
